+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/analysis/Layer.java
@@ -193,7 +193,19 @@ public enum Layer {
     /**
      * Kafka is a distributed streaming platform that is used publish and subscribe to streams of records.
      */
-    KAFKA(31, true);
+    KAFKA(31, true),
+
+    /**
+     * Pulsar is a distributed pub-sub messaging platform that provides high-performance, durable messaging.
+     * It is used to publish and subscribe to streams of records.
+     * Pulsar supports scalable and fault-tolerant messaging, making it suitable for use in distributed systems.
+     */
+    PULSAR(32, true),
+
+    /**
+     * A scalable, fault-tolerant, and low-latency storage service optimized for real-time workloads.
+     */
+    BOOKKEEPER(33, true);
 
     private final int value;
     /**