+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/management/ui/template/UITemplateInitializer.java
@@ -70,6 +70,8 @@ public class UITemplateInitializer {
         Layer.RABBITMQ.name(),
         Layer.MONGODB.name(),
         Layer.KAFKA.name(),
+        Layer.PULSAR.name(),
+        Layer.BOOKKEEPER.name(),
         "custom"
     };
     private final UITemplateManagementService uiTemplateManagementService;